Add files via upload
I have loved Esperanto ever since I first learned about it. It is incredibly easy to learn, and I decided I wanted to try and convince more people to give it a try. I think what pushed me the most was a couple weeks ago, I was on r/polyglot, one of the language learning subreddits, and someone asked if it was worth it to learn Esperanto. A lot of the comments were uninformed and negative, talking about how "no one speaks that" and that it was a waste of time. My goal here was to hopefully show that it most certainly is not a waste of time, and that there are some good benefits of learning it!
The main focus of this, for me, was exercising my learned skills in data scraping. I had SUCH a hard time learning how to do it this semester, and I still struggled with it for hours while I was working on this. So, I decided to challenge myself and do the thing I was worst at, for four different datasets on four different websites. I also challenged myself to go out of my comfort zone with my world map. That alone took a couple hours of research, and staring at examples and page after page of errors. I am really proud of the way it turned out.
I tried to keep the explanations to a minimum, and linked out to as many things as I could. I love this subject, so it was difficult to determine when I was giving too much information that it might take away from the project itself. I imagine that my audience is made up of people that know a bit about coding and what a dataset is, but don't necessarily know how to do things that I am doing. That's why I tried my best to explain what I was doing every step of the way, in comments and in plain text.

Thank you for your efforts this semester, and have a good summer!